Labels.py doc
